ai_service/requirements.txt
langgraph==0.2.45
langchain-core==0.3.27
langchain-mistralai==0.2.2
langchain-openai==0.2.14
pydantic==2.10.3
fastapi==0.115.6
uvicorn==0.34.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
python-dotenv==1.0.1
fastmcp
